debug, now OK with example input, next move to cluster

+++ core/src/main/scala/org/apache/spark/scheduler/Schedulable.scala
@@ -32,7 +32,7 @@ private[spark] trait Schedulable {
   // child queues
   def schedulableQueue: ConcurrentLinkedQueue[Schedulable]
   def schedulingMode: SchedulingMode
-  def weight: Int
+  def weight: Double
   def minShare: Int
   def runningTasks: Int
   def priority: Int
